kafka-proxy/requirements.txt
flask==1.0.3
kafka-python==2.0.2
waitress